requirements.txt
ccxt
python-dotenv
telethon
aiosqlite